Add files via upload
This is my ETL end-to-end architectural diagram. Here, I used Apache Airflow for workflow orchestration and Docker for containerization. We have two different options which are Databricks and Visual Studio Code,, but in my context, I chose Visual Studio Code, and later I will also test Databricks.  While, talking about the process we used, we have two different options for data source,e which are API based. However, I used Finnhub API for data sources. In the project, we are using streaming data with the help of Kafka producer, and for the bigdata I will take help of Apache Hadoop. For transformation, we are using PySpark, which is a powerful tool to play with big data and files like parquet to upload to the cloud. Load section, we have two different options which are using local storage or cloud side if we use local storage it doesn't support parquet file so we have to store it on CSV or JSON file format and for cloud section we will use parquet to upload on cloud like S3 and which we will connect with AWS other services like Athena and Redshift. To make it visualize and taking insights of this data we are going to use different visualization tools like POWER BI, Tableau.